[Refactoring] Move async refactorings into their own files
